+++ ubelt/util_path.py
@@ -335,7 +335,62 @@ def ensuredir(dpath, mode=0o1777, verbose=0, recreate=False):
     return dpath
 
 
-class TempDir(object):
+class ChDir:
+    """
+    Context manager that changes the current working directory and then
+    returns you to where you were.
+
+    Args:
+        dpath (str | PathLike | None):
+            The new directory to work in.
+            If None, then the context manager is disabled.
+
+    Example:
+        >>> import ubelt as ub
+        >>> dpath = ub.Path.appdir('ubelt/tests/chdir').ensuredir()
+        >>> dir1 = (dpath / 'dir1').ensuredir()
+        >>> dir2 = (dpath / 'dir2').ensuredir()
+        >>> with ChDir(dpath):
+        >>>     assert ub.Path.cwd() == dpath
+        >>>     # changes to the given directory, and then returns back
+        >>>     with ChDir(dir1):
+        >>>         assert ub.Path.cwd() == dir1
+        >>>         with ChDir(dir2):
+        >>>             assert ub.Path.cwd() == dir2
+        >>>             # changes inside the context manager will be reset
+        >>>             os.chdir(dpath)
+        >>>         assert ub.Path.cwd() == dir1
+        >>>     assert ub.Path.cwd() == dpath
+        >>>     with ChDir(dir1):
+        >>>         assert ub.Path.cwd() == dir1
+        >>>         with ChDir(None):
+        >>>             assert ub.Path.cwd() == dir1
+        >>>             # When disabled, the cwd does *not* reset at context exit
+        >>>             os.chdir(dir2)
+        >>>         assert ub.Path.cwd() == dir2
+        >>>         os.chdir(dir1)
+        >>>         # Dont change dirs, but reset to your cwd at context end
+        >>>         with ChDir('.'):
+        >>>             os.chdir(dir2)
+        >>>         assert ub.Path.cwd() == dir1
+        >>>     assert ub.Path.cwd() == dpath
+    """
+    def __init__(self, dpath):
+        self.context_dpath = dpath
+        self.orig_dpath = None
+
+    def __enter__(self):
+        if self.context_dpath is not None:
+            self.orig_dpath = os.getcwd()
+            os.chdir(self.context_dpath)
+        return self
+
+    def __exit__(self, a, b, c):
+        if self.context_dpath is not None:
+            os.chdir(self.orig_dpath)
+
+
+class TempDir:
     """
     Context for creating and cleaning up temporary directories.